benchmarks: implement bitcoin transaction environment
Our creation of witness data is clumsy and wrong (we don't embed any actual
program with a real CMR, and if we stick an annex it, it is unused by our
C environment constructor) but our Elements implementation has these same
issues. Ulimately the transaction environment is not that important because
all the transaction jets are implemented as lookups of cached data.